Update README.md
I like this plugin, it's just not entirely self-evident how to use its functionality.  I read your response in the issues section that did a pretty good job of explaining just that, so I took the liberty of including it here (with a few stylistic changes of my own) so that any future users that might be interested can find it a bit easier.